@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+	  <TargetFramework>net9.0</TargetFramework>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -36,7 +36,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Framework\Cache\Cache.MemoryCache\Cache.MemoryCache.csproj" />
+    <ProjectReference
+      Include="..\..\..\src\Framework\Cache\Cache.MemoryCache\Cache.MemoryCache.csproj" />
     <ProjectReference Include="..\..\..\src\Framework\Host\Host.Generic\Host.Generic.csproj" />
   </ItemGroup>
 
